SCB-2094 Move the DAO package
