+++ datasource/etcd/key_generator.go
