+++ datasource/etcd/key_generator_test.go
